+++ src/board/system76/common/include/board/battery.h
@@ -34,13 +34,13 @@ bool battery_set_start_threshold(uint8_t value);
 uint8_t battery_get_end_threshold(void);
 bool battery_set_end_threshold(uint8_t value);
 
-int battery_charger_configure(void);
+int16_t battery_charger_configure(void);
 void battery_event(void);
 void battery_reset(void);
 
 // Defined by charger/*.c
-int battery_charger_disable(void);
-int battery_charger_enable(void);
+int16_t battery_charger_disable(void);
+int16_t battery_charger_enable(void);
 void battery_charger_event(void);
 void battery_debug(void);
